--- src/main/kotlin/com/project/byeoldori/config/WeatherApiProperties.kt
@@ -1,10 +1,0 @@
-package com.project.byeoldori.config
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.stereotype.Component
-
-@Component
-@ConfigurationProperties(prefix = "weather.api")
-data class WeatherApiProperties(
-    var key: String = ""
-)
