+++ src/main/kotlin/com/project/byeoldori/forecast/config/WeatherApiProperties.kt
@@ -1,0 +1,10 @@
+package com.project.byeoldori.forecast.config
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.stereotype.Component
+
+@Component
+@ConfigurationProperties(prefix = "weather.api")
+data class WeatherApiProperties(
+    var key: String = ""
+)
